requirement.txt
Box2D                         2.3.1
certifi                       2018.4.16
chardet                       3.0.4
cycler                        0.10.0
future                        0.16.0
gym                           0.10.5
idna                          2.7
kiwisolver                    1.0.1
matplotlib                    2.2.2
neat-python                   0.92
numpy                         1.14.5
pip                           10.0.1
pyglet                        1.3.2
pyparsing                     2.2.0
python-dateutil               2.7.3
pytz                          2018.4
requests                      2.19.1
setuptools                    27.2.0
six                           1.11.0
tqdm                          4.23.4
urllib3                       1.23
WebOb                         1.8.2
wheel                         0.31.1